added in null checks for CF

+++ library/cloudformation/index.js
@@ -47,6 +47,8 @@ function CloudFormationWrapper({ accessKeyId, secretAccessKey, region }) {
   }
 
   async function createStack(templatePath, stackName, parameters=[], enableTerminationProtection) {
+    if (!templatePath || !stackName) {
+      throw new Error(`Must include templatePath (string) and stackName (string) for CloudFormation`);
     const params = {
       StackName: stackName,
       TemplateBody: fs.readFileSync(templatePath, 'utf-8'),
@@ -59,6 +61,9 @@ function CloudFormationWrapper({ accessKeyId, secretAccessKey, region }) {
   }
 
   async function deleteStack({ stackName, options={} } = {}) {
+    if (!stackName) {
+      throw new Error(`Must include stackName (string) for CloudFormation`)
+    }
     const deleteParams = { StackName: stackName };
     const resp = await cf.deleteStack(deleteParams).promise();
     const { waitToComplete, stdout } = paramTools.getOptions(options);
@@ -78,11 +83,17 @@ function CloudFormationWrapper({ accessKeyId, secretAccessKey, region }) {
   }
 
   async function getStackOutputs(stackName='') {
+    if (!stackName) {
+      throw new Error(`Must include stackName (string) for CloudFormation`)
+    }
     const { Stacks } = await cf.describeStacks({ StackName: stackName }).promise();
     return Stacks.pop();
   }
 
   async function stackExists (stackName='') {
+    if (!stackName) {
+      throw new Error(`Must include stackName (string) for CloudFormation`)
+    }
     try {
       const { Stacks } = await cf.describeStacks({ StackName: stackName }).promise();
       return Stacks.length > 0;
@@ -97,6 +108,9 @@ function CloudFormationWrapper({ accessKeyId, secretAccessKey, region }) {
   }
 
   async function updateStack(stackName, templatePath, parameters=[], stdout) {
+    if (!stackName || !templatePath) {
+      throw new Error(`Must include stackName (string) and templatesPath (string) for CloudFormation`)
+    }
     try {
       const updateParams = {
         StackName: stackName,
@@ -118,6 +132,9 @@ function CloudFormationWrapper({ accessKeyId, secretAccessKey, region }) {
 
 
   async function waitOnStackToStabilize(stackName, stdout, retry=0) {
+    if (!stackName) {
+      throw new Error(`Must include stackName (string) for CloudFormation`)
+    }
     const describeParams = { StackName: stackName };
     const WAIT_INTERVAL = 5000;
     if (retry < 240) {
